fix val & raw instruction  #13

+++ libs/Ancient.Runtime/instructions/raw.cs
@@ -11,6 +11,6 @@
 
         }
 
-        public override long Assembly() => (long) _value;
+        public override ulong Assembly() =>  _value;
     }
 }
